Rename 'What is dispute about' to 'dispute_type'
Following on from PR #30, this changes the naming of the "What
is your dispute about?" concept to the new agreed term of
"Dispute Type"/`dispute_type`.

A "dispute type" is the concrete dispute that a taxpayer has
wrt an HMRC decision/penalty/etc. It is only relevant for some
case types. Examples are "lateness" or "amount of tax owed".