Add Rails generator for tasks (#51)
This allows us to scaffold steps very easily. Given a task (whose
decision tree and super step controller must exist already), it
generates a controller+spec, form+spec, and view for a new step.
It also adds the step to the routes and to the task's decision tree.

Each of these files and changes will either assume sensible defaults,
or have TODO comments and raised errors where things need to be changed.