Move i18n of step headers into individual locale files
This moves the step headers ("Question 1 of 99") from `en.generic.step`
to the individual task locale files. This way, we no longer have to
specify the second number in each and every view template and can just
change it in one place if the total number of questions in a task
changes.