Refactor "back" link on step pages into helper
The tasks shouldn't be using different code to implement the "Back to
previous step" functionality (two have the back links in the views, one
sets an ivar in the controller). This harmonises the different tasks and
DRYs up the logic.

Also, the current logic in the details task fails when the user has
already submitted the form as the ivar only gets set in the `edit`
action, but not the `update` action.

- Add `#previous_step_path` method to step controllers that defines the
  path for the previous path's edit action
- Add a `_step_header` partial for the step header markup
- Add `ApplicationHelper#step_header` that renders the step header
  partial with appropriate content
- Update all step views to use the helper