Override choices for ChallengedDecision form
We were using the default choices from the GOV.UK form builder for the
ChallengedDecision form, which implicitly work for yes/no radio buttons
but break when you navigate back to the page (the previously selected
item is no longer selected).

This changes the form to use `true`/`false` instead of `:yes`/`:no` and
also brings it in line with the other forms.