Create new cases in GLiMR.
This is a starting point as we don't have yet all the information to be sent to GLiMR
and also we don't have yet the summary page that will trigger the call to create the cases.

Other missing bits are the API call to store the TC number and confirmation code and return
the payment URL to redirect the user to.