Check your answers step (summary).
This PR implements most of the summary page based on the information we already have.
It will require continue to grow the page as we add more questions and journeys in the
datacapture app.

On submit it will hit the cases controller which will send the case to GLiMR and the
payments app and get back a URL to redirect the user for payment.