Documents upload bugfixes (#80)
* Fix a bug when deleting grounds for appeal documents with spaces.

If the grounds for appeal uploaded document contained spaces, when trying to delete it,
despite working fine, it didn't nil the filename in the tribunal case object, giving
the impression the document was still there and the delete didn't work.

* Fix a few CSS images not showing on prod.

* Small refactor to enable the possibility of returning false from persist! method.

Previously the `true` return value was hardcoded. This fixes a bug where the
grounds for appeal document upload might fail but still progress to the next
step and save the name of a (non-existing) file in the tribunal case record.