Generate and download case details PDF (#92)
* WIP - Generate and download case details PDF.

This is a rough first stage but it is working nicely. There is a need to
adapt the `DocumentUpload` class to be able to handle File/Tempfile inputs and
not only `ActionDispatch::Http::UploadedFile` as it is now.

* Some refactoring and other changes.

Preparing for the filename to be dynamic with the TC number once we store it in the
database, as now we don't have it. Also preparing for integration with the case
submission and step.