Implement the case details PDF design and real data.
As part of this task I've needed to do a bit of refactor and some other changes,
specially relating to presenters. This simplifies working with the PDF generator and
now for example we can use the same presenters we use for `check your answers` but
with a custom set of i18n so we have full flexibility.

Appending .pdf to the `check your answers` step will generate and render the PDF.

Other changes are the inclusion of a new tribunal case attribute `case_reference`,
storing it after succesfully calling GLiMR and payment services, and using it in
the generated and uploaded PDF.