Simplifying the maintenance of 2 duplicated sets of translations.
This commit avoid the need to have 2 completes sets of translations to be used
in the `check your answers` HTML view and PDF view, as using i18n cascading will
allow us to only define those translations that are different, and the ones that
are identical will be used.

A convenience helper method `pdf_t` has been exposed to the view to avoid lots of
repetition and work out the offset automatically.