Implement the case details PDF design and real data. (#99)
* Implement the case details PDF design and real data.

As part of this task I've needed to do a bit of refactor and some other changes,
specially relating to presenters. This simplifies working with the PDF generator and
now for example we can use the same presenters we use for `check your answers` but
with a custom set of i18n so we have full flexibility.

Appending .pdf to the `check your answers` step will generate and render the PDF.

Other changes are the inclusion of a new tribunal case attribute `case_reference`,
storing it after succesfully calling GLiMR and payment services, and using it in
the generated and uploaded PDF.

* Simplifying the maintenance of 2 duplicated sets of translations.

This commit avoid the need to have 2 completes sets of translations to be used
in the `check your answers` HTML view and PDF view, as using i18n cascading will
allow us to only define those translations that are different, and the ones that
are identical will be used.

A convenience helper method `pdf_t` has been exposed to the view to avoid lots of
repetition and work out the offset automatically.