Introduce a new Home page and move Task list to a separate one.
This PR introduces as per prototype design a new Home page with links to different
actions a given user can do (appeal, close and pay), with the first link (appeal)
linking to what until now was the Home page.

Copy changes/markup for the Task list to mimic the prototype.