Rename penalty_amount DB attribute to penalty_level. (#103)
This is a first PR needed for a further introduction of a new `penalty_amount` attribute,
which is going to be free text and will be filled by the user, in addition to choosing
the penalty level as it is now. This the renaming of the current `penalty_amount` attribute
to something more meaninful.