Optionally ask for a penalty amount (free text) depending on radio buttons.
This PR is a continuation to #103 and introduces a new DB attribute, `penalty_amount`
which will allow us to store a free text given by a user.

The input field will show or hide according to the prototype when selecting specific
radio buttons, and because of this, some custom javascript was written to handle it,
as govuk elements form builder wouldn't allow us to have this specific behaviour.

The javascript should be generic enough to be used for any other step behaving in a
similar way.