Appealed indirect tax flow (#110)
* Challenged decision status flow.

This is the basic workflow for an indirect tax that requires the
appellant to have appealed the decision to HMRC, first.

* Improve spec coverage

Putting the cart before the horse a bit.  This is because Jesus took me
through a quick walkthrough while pairing and so we didn't start with
the tests.  Retrofitting them no

* Refactor spec to make it easier to maintain

This has some boilerplate in it.  However, I think this aids the overall
clairity, where shared examples would likely obscure them somewhat.