Replace TaxpayerType with more generic value object
This is the first step towards generalising the appellant/
representative flow. As the "entity type" (individual/company/other) is
the same across appellants and representatives, they should be
represented by a single value object.

Also brings the copy and set of choices in line with the prototype (but
the taxpayer type form is about to be replaced).