Refactor common logic out of taxpayer details forms
As another step towards reusable forms for repellants, this groups the
common fields between different kinds of taxpayer (individual/company)
into a common superclass.

It also adopts a new naming scheme using "organisation" instead of
"company" for the form objects (not the controllers yet, because they
will need to be part of a larger change towards the repellant flow).