Add ability to specify a free text 'other' case type.
When the user selects `none of the above` option in the case type list, a new
input text is shown for the user to specify their tax or dispute type.

Updated decision trees and also a bit of refactor in the cost/fee presenters as
many of the questions/answers are common, just in a slighly different order.
It might be that we end up with the exact same order once the prototype is finished,
in which case, it will make more sense to just have a single presenter, but for now
I think it is safer to leave it as 2 separate presenters to have more flexibility.