Rename 'company' fields to 'organisation'
Contactable entities are either individuals, companies, or "other
organisations". The latter two are treated identically within the
application, it's just a separate choice in the repellant flow for
UX reasons. For the sake of good naming, we'll use the term
'organisation' throughout the code because that semantically covers
both companies and "other organisations".

- Make details decision tree work properly with the "other organisation"
  option
- Rename all 'company_' fields to 'organisation_'
- Stub the taxpayer type where appropriate in specs (instead of passing in
  a specific `ContactableEntityType`)