Make the paths more adaptable
While many of the steps are the same between the `Fee Free Service` and
the `With The Fees` version, the `step` paths did not clearly reflect
this.  They were all `step/cost`.

After discussions, we agreed that `step/appeal` was a more appropriate
path as it works with either version without needing to be altered.

This (rather hefty) commit changes all sensible occurrences of
`step/cost` to `step/appeal`.  `Cost` as a class and/or module name
still remains in a couple of places as it was nonsensical to change
these.