Store the current step path in session for a later user.
This improves over the previous problem of knowing where to redirect back the user
when uploading or removing files with JS disabled, not needing to submit a hidden
input with the path.

At the moment only steps doing document uploads are using this session value but
it seemed sensible to have this set in all StepController children, mainly for
consistency, and because we might need it for other situations.