Several positioning, non-JS enhancements, error handling and copy. (#157)
* Several positioning, non-JS enhancements, error handling and copy.

This PR will position the letter checkboxes (appeal flow) above the dropzone area
and will work for any browser, granting there is JS available. For non-JS browsers
the checkboxes will remain below the dropzone area, but will also be functional.

In order to do this, the HTML5 `form` attribute is being used in the checkboxes, but
unfortunately this attribute is not supported by IE (none version) so we put in place
a polyfill JS workaround for this. Non-JS browsers, regardless if it is IE, will work
as described previously.

This PR will also change some error copy messages, and validate file size and file type
in the client size if JS available, otherwise backend will be used.

Some other markups changes, styling and improvements. More styling will be required.

* Some improvements.