Closure journey `Check your answers` page, presenters and PDF.
This PR will add the `Check your answers` page to the Closure flow. A bit of refactor
was done in the presenters, and a new one for Closure has been created. Also some minor
changes were needed in the PDF creator service object to be able to handle the two
different `answer` pages we have now (appeal and closure).

Please note submitting will not work yet, that will be part of a follow up PR, but better
to get this merged as otherwise it could cause conflicts if delayed too much.