Appeal and Closure submission and confirmation pages.
Refactored a bit the cases_controller as we have now two submissions with some
small differences.

It looks like there is a bit of duplication in the views/i18n, but I decided not
to do an early abstraction as I think the confirmation pages might need to change
their markup or copy later on, and be different for appeals or closures.

The survey links will be introduced as part of another PR.