Navigation stack for "back" navigation in steps
Instead of the current complex "back" button logic, where we need to
figure out for each step what the previous step would be given the state
of the `TribunalCase`, add a navigation stack that keeps track of the
path the user took through the steps to arrive at the current step.

This does not yet include the removal of the old logic (will be done
separately to get this change in ASAP).