Change the order of the case_type and challenged_decision steps. (#180)
* Change the order of the case_type and challenged_decision steps.

Also, to be able to mimic the new decision tree more easily, a new `ask_challenged` attr
has been added to the CaseType value objects so we can trigger the challenged step for the
case types that need it.

We need to audit all the case types to make sure all the possible combinations are covered
and matches the decision tree. Also the `challenged_status` step need to change as right now
is pretty much a placeholder. This will be made as separate PRs.

* Some improvements.