Add back links to kickout pages. (#186)
I could only find 2 kickout pages so far. We might need some more later on. Unless we also
want to add these back links to the confirmation pages but it doesn't make too much sense to me.

A new helper method introduced for readability, but the only thing it does is call the existing
method `step_header` with nil values. We render the same partial but without the progress copy.