Mark a case as submitted and disallow users re-submitting. (#189)
As part of this PR, a bit of refactor was done to check whether we have a tribunal case
in first place, and also if it is in a valid status. Some of this checks are not needed in
all actions or controllers so this is the reason why they are 2 separate checks.

We still continue allowing the user to generate the PDF (if a valid session exists) and
also we now consider the `start` controllers the beggining of the flow and these will
create if needed the tribunal case DB entry, instead of the `case type` controler as before.
This fixes a small issue with the back links taking the user from the `case type` step to
the home page, instead of to the `appeal start page` or `closure start page`.

Another PR will follow to implement the commented out before_actions.