Add before_action validations to controllers not inheriting from StepController (#190)
This PR is a follow up to the previous one to add the before_filter validations to some
controllers not covered by the StepController superclass.

Although there is a bit of repetition, this ensures the code is more readable and we know
always which controllers implement which filters. In case we were to have more non-step
controllers in the future, a better solution might be inheritance or a concern, but
currently we only have 2 of these non-step controllers.