Handling of unexpected exceptions. (#191)
Show the user a nice error page (design/copy TBC) but also report the exception to Sentry.
We are already handling case not found in session/session expired and case already submitted
with a similar approach, but in these scenarios we will not report to Sentry as this is a
recoverable user-error. We can decide otherwise later.