Integration with GOV.UK Notify.
This will use the `govuk_notify_rails` that acts as the connector for `notifications-ruby-client`
via a custom ActionMailer delivery method.

An example of email has been implemented (`case_submitted`) but the template is a
placeholder and we need to change it. Other emails can be created in a very similar way.