Ensure we send email to the correct party
The notification should go to the party that fills out the request;
either the taxpayer or their representative.  These changes make the
contact email associated with `tribunal_case.user_type` mandatory.  In
the current configuration of the app, this *should* be the email
associated with the first set of contact details.

Emails entered subsequently that do not belong to the same
`user_type` (`taxpayer` or `representative`) are not required.