Kickout page for users having problems uploading documents.
This page will allow to download/print the case PDF but will hide previously uploaded
documents, if any. Copy TBC.

When clicking Finish button, the session will be reset and the user will be redirected
to the survey page, as we do with the confirmation page.

This PR is only for the appeal journey. A similar PR will be created for the closure.