Kickout page for users having problems uploading documents (closure). (#213)
This page will allow to download/print the case PDF but will hide previously uploaded
documents, if any. Copy TBC.

It works in the same way as in the Appeal flow, and some of the code can be seen as
duplication, but IMO trying to avoid this small duplication will result in a more
confusing code.

Also as part of this PR, two DB fields were removed as were redundant.