Email if GLiMR is down or there is a failure. (#230)
* Refactor to trap errors

Right now it traps and logs everything.  This will change shortly.  It
is groundwork for sending an email to the tribunal if the case is down.

It also corrects a minor false-negative spec issue (`.call` vs
`.call!`, both of which raised the spec'd exception).

* Change case status on Glimr error

Log and re-raise all other errors.

* Some refactor of Todd's previous work, and implementing the email.