Dynamic personalisation for case confirmation email.
Depending on who is the recipient (taxpayer/representative), the type of case
(application or appeal) and if we have or not case reference, some of the
personalisation that we send to Notify will change.

As potentially we will need this in other emails, a little presenter to encapsulate
this logic has been created.

Also as part of this ticket a bug was found with `equal?` not returning the expected
result for `UserType`. This has now been changed to `eql?` and tests adapted.