Smoketests (#241)
* Enable happy path smoke test(s)

Right now it only does the simplest income tax happy path.

* Make the origin of the request explicit

Right now, it does not submit after 'Check your answers'.  However, it
is likely that it will do in a later commit. This change ensures that if
it *does* get run against production, the Tribunal will be able to see
that the request originated from MoJ Digital Smoketests.

* Ensure case submission works

And that we get back a TC number.