Enable smoketesting from Jenkins (#249)
* Enable smoketesting from Jenkins

At the time of writing, Jenkins uses an outdated version of ruby. This
makes it difficult to install some of the gems required for the smoke
tests.  It also lacks PhantomJS, as well as the dependencies needed to
build it from scratch (bzip).

However, building a docker container is something it already does in a
well-understood way.  This push builds a container, which then runs all
the smoke tests.

Note: I experimented with skipping the development dependencies, but
decided against as this required a more complex `Dockerfile`, but did
not significantly reduce the container build time.

* Add CI script example to README

This is the simplest possible example.

* Update README.md

* Update README.md