Rescue and send to Sentry unexpected exceptions on mailers. (#254)
This will ensure we are notified about any possible exceptions happening on the
NotifyMailer, i.e. missing personalisation, template not found, etc.

We will append extra context to the Raven event to know which template triggered the
exception and if any, the personalisation sent (some details filtered out for privacy).