'Check your answers' has Grounds for Appeal text and document (if any attached).
Ensure we show all the relevant information for Grounds for Appeal. If the user enters both
(free text and attach a document) we show both. If only one of these is filled, we will only
show the entered information.

The same approach will be follow for the `hardship reasons` and other steps that contain
text and/or document attachment, in follow-up PR's.