Re-work the closure `having problems uploading` decision tree.
We were using the same logic as in the Appeal journey, but for Closure, having problems
uploading doesn't mean the user can't submit their case, so we are now continuing to the
`check your answers` page as usual.

Also, as part of this, a bit of layout changes have been made (both html and pdf versions)
so the problems typed by the user are shown in the `documents submitted` section, instead of
the list of documents.
These changes are for Appeal (pdf only) and Closure (pdf and html).

https://www.pivotaltracker.com/story/show/141243869