JS-dependant solution for dropzone duplicate file names.
This is a spike JS solution, working with Dropzone events, to detect duplicate already
uploaded documents, and if any found, rename on the fly the filename appendind (1) until
a unique filename is obtained. This new filename gets uploaded and added to the list of
uploaded files, transparent to the user.

This solution is simple and works making use of the hook provided by Dropzone `renameFilename`
but will not have any effect on the non-JS uploader version.

Another backend solution will be spiked now. Do not merge.