Never send unsanitized input to 3rd party APIs
While the datacapture app is protected via standard Rails protocols, we
cannot ensure our partners always will be and so we want to be careful
about what information we forward to them.  To this end, this PR scrubs
html tags, encodes characters that are problematic in HTML, and
transforms or removes a number of characters and statements that are
commonly used in SQL injection attacks.

This is *not* comprehensive prevention.  Rather, it is intened to ensure
that we do not accidentally propogate the most common attack vectors.