Non-JS solution for dropzone duplicate file names.
Detect duplicate already uploaded documents, and if any found, rename on the fly the
filename appendind `(n)` where `n` is an incrementing number starting in 1, until a
unique filename is obtained. This new filename gets uploaded and added to the list of
uploaded files, transparent to the user.

Possible combinations:

- Existing `image.jpg`, uploading `image.jpg`, renamed to `image(1).jpg`
- Existing `image.jpg` and `image(1).jpg`, uploading `image.jpg`, renamed to `image(2).jpg`
- Existing `image(1).jpg`, uploading `image(1).jpg`, renamed to `image(1)(1).jpg`

Also, as part of this PR, we are now returning date-sorted files from the S3 bucket.