Cleaning up uploader error translations.
It turns out we can simplify this by adding the errors to the default.yml so Rails will take care
automatically of this.

The long version of the error (the one with interpolation) continue to be in `document_upload.yml`
as this one will not be handled automagically by Rails and we need to do the interpolation ourselves.