Prevent users submitting the same case multiple times.
This PR introduces an intermediate case status `in_progress` which is set as soon as
the user clicks submits, while the submission to GLiMR is being processed, and we reset
the status back to `submitted` once we get the result from GLiMR.

This stops the user from clicking quickly the submit button, or submitting and going back
and forward, and the case will not be duplicated. The user will get the message about their
case being already submitted, and we will show the case reference if any, and the link to
download the PDF, as usual.

https://www.pivotaltracker.com/story/show/141857773

Another PR will also introduce a frontend improvement to give the user feedback once clicking
the submit button, and disable or hide it to avoid also duplications, but this will be a
JS-dependant solution so this PR is a catch-all solution for the odd edge case.