Add explanatory copy when TC number is not present.
Add a explanatory copy in the confirmation page and the email sent to the taxpayer in case
the TC number is not present, to tell them it will be sent at a later stage.

Only the 'dev/test' Notify environment has been updated with the new template variables, so
production is not affected, but once we deploy to PROD, we need to ensure to update the
corresponding template too, otherwise as we are sending different personalisation, it will fail.

https://www.pivotaltracker.com/story/show/141877681