Fix for `check your answers` not showing representative details sometimes.
A condition exists where filling the form as a representative, instead of as a taxpayer
having a representative, will not set the attribute `has_representative` and thus, even
if the details are filled, we don't show them in the `check your answers` page.

With this small change we ensure that regardless of the path the user follows to fill the
form, if they are asked for representative details, we set the `has_representative` to YES.
This will catch both scenarios just fine.

https://www.pivotaltracker.com/story/show/141957833