Tailor challenge copy for direct/indirect taxes.
Depending on the case type being direct or indirect tax, we need to show a slighly different
copy in the challenge step, but also in the check your answers.

The solution is not pretty but probably not worth overengineering it, specially as there is a
big PR to change the check your answers presenters waiting to be merge and there will be conflicts.

https://www.pivotaltracker.com/story/show/140878591