Some kickout pages should not redirect to the survey.
As per Josh request, some kickout pages should not redirect the user to the survey when
they click the 'Finish' or 'Start again' button, as those kickout are too early into the
journey or are errors and will not provide any meaninful feedback.

We continue to have the survey in some other kickout pages, like the `having problems uploading
documents` as this one is almost to the end of the journey.

https://www.pivotaltracker.com/story/show/142587175