Remove PDF-specific check answers translations
We originally added the ability to override specific questions/answers
in the "Check your answers" PDF to make them different from the copy
shown on the HTML version. After conversations with design and content,
this is no longer required - so this removes both the translation and
the underlying code from the app.

- Remove `UploadHelper` and `UploadHelper#pdf_t`
- Use regular `I18n.translate`/`t` instead of `pdf_t` in PDF view
  templates
- Remove remaining PDF-specific translations