Restoration case 'decision challenge' to list indirect options.
This change will take care of the `decision challenge` for restoration cases, where:

- when asking YES, the list of options must behave as an indirect tax.
- when asking NO, there should be a kickout like for direct tax.

Also, we are not asking for disputes and we jump directly to lateness.

As part of another PR we need to revisit the 2 possible kickout pages to change the copy
depending if it is a direct or indirect tax (as we did for the header of the challenge question).

https://www.pivotaltracker.com/story/show/142498857