Prevent mutation tests from slowing to a crawl
The final mutation for the mapping code determiner still takes a long
time to complete, but we're talking maybe ten more seconds, not three
more hours.

I *suspect* it has something to do with object freezing as the
changed specs around `#appeal_or_application` complain about the value
object doubles being (unexpectedly) frozen if you run them in isolation.

For now, the workaround for `#mapping_code` is to avoid instantiating it
with the calling `TribunalCase` instance as an argument; pass the
instance in afterwards as per the changed models/specs. I imagine this
will apply to any future proxy object call as well.

The fix for `#appeal_or_application` was more straightforward.  Aside
from the aforementioned warnings about the frozen doubles, the solution
was simply to change real objects to doubles and refactor the specs to
follow the code a bit more closely.

I also enabled slow spec reporting.  This revealed some things we may
want to address later, but turned out not to be relevant for this
particular issue.