Branching of challenge kickout pages for direct/indirect/restoration cases.
This PR will implement the neccessary logic to branch the kickout challenge pages depending on
the kind of tax (direct/indirect) and the snowflake `restoration case`.

Also as part of this PR, we lay out the path to a slight refactor to group together all the challenge
logic and decision in their own decision tree class.
There are several TODOs which is going to be the next PR.

https://www.pivotaltracker.com/story/show/142498857