Fix issues with "I'm having trouble uploading docs"
When a user ticks the "I'm having trouble uploading documents" checkbox
on the multi-upload page, then submits, then goes back to the page
later, they cannot undo that choice because the validation will always
fail.

The documents submitted are checked in a validation _before_ the new
value for "I'm having trouble" is saved - but `TribunalCase#documents`
returns no documents when "I'm having trouble" has previously been saved
as true. So the validation fails, the fields don't get reset, and
everyone is sad.

- Rename 'having trouble uploading' fields to be more clear
- Clear (or do not save in the first place) explanation field if
  'having trouble uploading' isn't ticked
- Ensure `having_problems_uploading` is set on `TribunalCase` **before**
  validations are run