Eliminate mutant speed and memory issues
Mutant was struggling with the `self` declaration that `TribunalCase`
was passing to the `MappingCodeDeterminer.new` method.  I suspect the
issue was caused by the 17 `.has_value_object` declarations that open
any new instance of `TribunalCase`; each, in turn, instantiates a value
object with a minumum of two values.  I believe this caused mutant to
build an exceptionally large abstract syntax tree, which then has to be
traversed multiple times.

After attempting multiple different fixes with limited success, I
finally settled on changing the determiner to a mixin.  CS suggested
that the `TribunalCase` spec could then be reduced to a simple check
to ensure the mixin was present, while the `MappingCodeDeterminer` spec
was easily modified by passing in a dummy class as the subject.