Kill hulking mutants (#343)
* Prevent mutation tests from slowing to a crawl

The final mutation for the mapping code determiner still takes a long
time to complete, but we're talking maybe ten more seconds, not three
more hours.

I *suspect* it has something to do with object freezing as the
changed specs around `#appeal_or_application` complain about the value
object doubles being (unexpectedly) frozen if you run them in isolation.

For now, the workaround for `#mapping_code` is to avoid instantiating it
with the calling `TribunalCase` instance as an argument; pass the
instance in afterwards as per the changed models/specs. I imagine this
will apply to any future proxy object call as well.

The fix for `#appeal_or_application` was more straightforward.  Aside
from the aforementioned warnings about the frozen doubles, the solution
was simply to change real objects to doubles and refactor the specs to
follow the code a bit more closely.

I also enabled slow spec reporting.  This revealed some things we may
want to address later, but turned out not to be relevant for this
particular issue.

* Eliminate mutant speed and memory issues

Mutant was struggling with the `self` declaration that `TribunalCase`
was passing to the `MappingCodeDeterminer.new` method.  I suspect the
issue was caused by the 17 `.has_value_object` declarations that open
any new instance of `TribunalCase`; each, in turn, instantiates a value
object with a minumum of two values.  I believe this caused mutant to
build an exceptionally large abstract syntax tree, which then has to be
traversed multiple times.

After attempting multiple different fixes with limited success, I
finally settled on changing the determiner to a mixin.  CS suggested
that the `TribunalCase` spec could then be reduced to a simple check
to ensure the mixin was present, while the `MappingCodeDeterminer` spec
was easily modified by passing in a dummy class as the subject.

* See if circle version 1 fixes the occassional hangs