Email reminders before a case is purged from database.
This PR prepares for sending email reminders (2 at the moment, easy to extend/change)
given a set of rules, to notify the user their case is about to be purge if they don't
complete and submit soon.

There are still some unknowns (email copy and thus personalisation, who will be the
recipient of the email: the account email address, or the case email address...) and
also, the rake tasks are commented out so code is not being executed at this time.

First reminder:
https://www.pivotaltracker.com/story/show/143109753

Second reminder:
https://www.pivotaltracker.com/story/show/143109867