Test for presence of SAVE_AND_RETURN_ENABLED env var
Template deploy converts 'true' in the deploy repo to "True"
when the env var is read, by the ruby code. This commit means
any value found for SAVE_AND_RETURN_ENABLED (including "false")
will turn on the feature. This means, to disable the feature,
we need to ensure the env var is unset, not set to false.