Delete action in cases controller.
Not really useful at the moment, but eventually this will be integrated
in the cases portfolio and user would be able to delete their cases.

Also, not ensuring right now the permissions to delete a given case by UUID,
but once the relationship between logged-in user and cases exists, we might
want to check this, although UUIDs might be safe enough not to be guessed.

https://www.pivotaltracker.com/story/show/143051147