Purge old records (#376)
* Make the spec easier to read

I originally put the shared examples at the top of the only spec they
related to.  As I've continued to work with the spec, however, I've
found that placement to be more annyoning that helpful. As a result, I'm
moving them out to their own shared examples file.

* Enable expired tribunal_case records to be removed

Ultimately, this will be plugged into the daily cron job, but can also
be run independently.

* default to 14 days for case expiry

It's what we talked about in the team, so it's one less env var to set