New sign up page for Save&Return
This replaces the dummy sign up page for save and return with a real
sign up page that creates a user in the database (but doesn't associate
it with a case yet).

Because of the extra data we need to capture as part of the sign up
process (a reference and email confirmation), this needs to go outside
of the default Devise sign up process, and introduces two form objects
which work much like they do in the step forms in the rest of the app.