Add relationship between User and TribunalCase.
This add the foreign key to the `tribunal_cases` table, and prepares the way to
be able to return all cases for a given user.

The relationship between `tribunal_case` and `user` is optional as we will not
have a user to associate to the case until the user actually creates an account.

This PR is part of the big ticket:
https://www.pivotaltracker.com/story/show/143050769