Cases portfolio (list of user's cases), and authentication.
This PR implements the `index` action, which will show the portfolio view with
a list of current logged-in user cases and some basic information (still WIP).

The `delete` button is also implemented, and now only cases owned by the
currently logged-in user can be deleted.

As part of this PR, both actions (#index and #create) are authenticated (Devise).

https://www.pivotaltracker.com/story/show/143050769