Ability to resume incomplete cases from the portfolio.
This PR implements the `resume` action, which will (assuming correct authentication)
assign the choosen tribunal case to the current session and redirect to the corresponding
`check your answers` page, depending on the case intent.

Also, the list of cases in the portfolio has been limited to 'pending' i.e. not submitted.

https://www.pivotaltracker.com/story/show/143050769