Refactor login controller and form objects to be more Devise-friendly.
Now we don't need to provide the `user case reference` in the login form, it is much more
simple to let Devise handle a lot of the logic and just inject a bit of code to link the
current tribunal case in session (if any) with the user.

This also handle nicely the return path after login, the different login errors, and
the DB magic columns (some of them we don't need, so using virtual attrs).