Password resets, logout, change in flows, and more.
This PR introduce the password resets, the logout, and all the confirmation pages
for these, and for saving a case depending if it is on sign-up or on sign-in.

Also the user menu in the header is now mostly functional (except the change password).

Some pending work for extracting to a service object the handling of the linking of the
user to the tribunal case, and for sending the email (at the moment the email is being
sent only on account creation, not on existing account).

Copy need to be reviewed everywhere. At the moment, just sensible defaults.

Work in this and some previous PR relate to:

* https://www.pivotaltracker.com/story/show/143051113
* https://www.pivotaltracker.com/story/show/143051055
* https://www.pivotaltracker.com/story/show/143051095
* https://www.pivotaltracker.com/story/show/143050707