Resume `check your answers` page.
This PR ensure we only show in the check your answers page those sections
that have at least an answer and avoid having missing translation exceptions.

Also it renders a different template for the `resume` action so we can get rid of
the declaration, modify the title and also we will have (WIP) a button to
resume the case, which will redirect the user to the last answered step.