Implement 'continue/resume' a previously saved case.
This PR implements the missing functionality to continue a saved case in the point
where the users left. It can be triggered from the resume `check your answers` page,
or in the scenario where the user didn't answer anything yet, we direct them to the
`case_type` step, as at the moment this is the first step.

Two helper routes points to the first step of the appeal or closure flows, so it is
easy to change in the future in case we wanted to redirect the users to a different url.