Copy and layour changes to update password screen.
As per Jason design. Including a kind of hack to force validation when new password
is left blank, as weirdly enough Devise will not flag this as an error and will just
ignore silently the update, but return true as if it was updated, which is very confusing
for the user (they would not get a confirmation email though).

To bypass this, when the new password is left blank, we replace the empty string with a
single character to force the superclass method trigger a validation error on the field.

https://www.pivotaltracker.com/story/show/143048897